# Update TimeFunction to avoid FIPS self check time affecting results