--- a/tool/speed.cc
+++ b/tool/speed.cc
@@ -163,6 +163,13 @@ static uint64_t g_timeout_seconds = 1;
 static std::vector<size_t> g_chunk_lengths = {16, 256, 1350, 8192, 16384};
 
 static bool TimeFunction(TimeResults *results, std::function<bool()> func) {
+#if defined(BORINGSSL_FIPS)
+  // The first time |func| is called an expensive self checks might run that
+  // will skew the iterations between checks calculation
+  if (!func()) {
+    return false;
+  }
+#endif
   // total_us is the total amount of time that we'll aim to measure a function
   // for.
   const uint64_t total_us = g_timeout_seconds * 1000000;
@@ -187,6 +194,9 @@ static bool TimeFunction(TimeResults *results, std::function<bool()> func) {
     }
   }
 
+  // Don't include the time taken to run |func| to calculate
+  // |iterations_between_time_checks|
+  start = time_now();
   for (;;) {
     for (unsigned i = 0; i < iterations_between_time_checks; i++) {
       if (!func()) {